Moved database model to its own file
